requirements.txt
matplotlib
numpy
pillow
scikit-learn
torch
torchvision
tqdm
